+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaSessionPermissionTest.java
@@ -27,7 +27,6 @@ import static androidx.media3.common.Player.COMMAND_SET_MEDIA_ITEMS_METADATA;
 import static androidx.media3.common.Player.COMMAND_SET_TRACK_SELECTION_PARAMETERS;
 import static androidx.media3.session.MediaUtils.createPlayerCommandsWith;
 import static androidx.media3.session.MediaUtils.createPlayerCommandsWithout;
-import static androidx.media3.session.SessionCommand.COMMAND_CODE_SESSION_SET_MEDIA_URI;
 import static androidx.media3.session.SessionCommand.COMMAND_CODE_SESSION_SET_RATING;
 import static androidx.media3.session.SessionResult.RESULT_SUCCESS;
 import static androidx.media3.test.session.common.CommonConstants.SUPPORT_APP_PACKAGE_NAME;
@@ -37,7 +36,6 @@ import static com.google.common.truth.Truth.assertThat;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
 import android.text.TextUtils;
 import androidx.media3.common.MediaItem;
@@ -51,7 +49,6 @@ import androidx.media3.test.session.common.HandlerThreadTestRule;
 import androidx.media3.test.session.common.MainLooperTestRule;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-import androidx.test.ext.truth.os.BundleSubject;
 import androidx.test.filters.LargeTest;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.Futures;
@@ -185,31 +182,6 @@ public class MediaSessionPermissionTest {
   }
 
   @Test
-  public void setMediaUri() throws Exception {
-    Uri uri = Uri.parse("media://uri");
-    createSession(
-        createSessionCommandsWith(new SessionCommand(COMMAND_CODE_SESSION_SET_MEDIA_URI)),
-        Player.Commands.EMPTY);
-    controllerTestRule
-        .createRemoteController(session.getToken())
-        .setMediaUri(uri, /* extras= */ Bundle.EMPTY);
-
-    assertThat(callback.countDownLatch.await(TIMEOUT_MS, MILLISECONDS)).isTrue();
-    assertThat(callback.onSetMediaUriCalled).isTrue();
-    assertThat(callback.uri).isEqualTo(uri);
-    BundleSubject.assertThat(callback.extras).isEmpty();
-
-    createSession(
-        createSessionCommandsWith(new SessionCommand(COMMAND_CODE_SESSION_SET_RATING)),
-        Player.Commands.EMPTY);
-    controllerTestRule
-        .createRemoteController(session.getToken())
-        .setMediaUri(uri, /* extras= */ Bundle.EMPTY);
-    assertThat(callback.countDownLatch.await(NO_RESPONSE_TIMEOUT_MS, MILLISECONDS)).isFalse();
-    assertThat(callback.onSetMediaUriCalled).isFalse();
-  }
-
-  @Test
   public void setRating() throws Exception {
     String mediaId = "testSetRating";
     Rating rating = new StarRating(5, 3.5f);
@@ -223,9 +195,7 @@ public class MediaSessionPermissionTest {
     assertThat(callback.mediaId).isEqualTo(mediaId);
     assertThat(callback.rating).isEqualTo(rating);
 
-    createSession(
-        createSessionCommandsWith(new SessionCommand(COMMAND_CODE_SESSION_SET_MEDIA_URI)),
-        Player.Commands.EMPTY);
+    createSession(SessionCommands.EMPTY, Player.Commands.EMPTY);
     controllerTestRule.createRemoteController(session.getToken()).setRating(mediaId, rating);
     assertThat(callback.countDownLatch.await(NO_RESPONSE_TIMEOUT_MS, MILLISECONDS)).isFalse();
     assertThat(callback.onSetRatingCalled).isFalse();
@@ -248,9 +218,7 @@ public class MediaSessionPermissionTest {
 
     // Change allowed commands.
     session.setAvailableCommands(
-        getTestControllerInfo(),
-        createSessionCommandsWith(new SessionCommand(COMMAND_CODE_SESSION_SET_MEDIA_URI)),
-        Player.Commands.EMPTY);
+        getTestControllerInfo(), SessionCommands.EMPTY, Player.Commands.EMPTY);
 
     controller.setRating(mediaId, rating);
     assertThat(callback.countDownLatch.await(NO_RESPONSE_TIMEOUT_MS, MILLISECONDS)).isFalse();
@@ -315,12 +283,10 @@ public class MediaSessionPermissionTest {
 
     public @Player.Command int command;
     public String mediaId;
-    public Uri uri;
     public Bundle extras;
     public Rating rating;
 
     public boolean onCommandRequestCalled;
-    public boolean onSetMediaUriCalled;
     public boolean onSetRatingCalled;
 
     public MySessionCallback() {
@@ -333,7 +299,6 @@ public class MediaSessionPermissionTest {
       mediaId = null;
 
       onCommandRequestCalled = false;
-      onSetMediaUriCalled = false;
       onSetRatingCalled = false;
     }
 
@@ -345,17 +310,6 @@ public class MediaSessionPermissionTest {
       this.command = command;
       countDownLatch.countDown();
       return MediaSession.Callback.super.onPlayerCommandRequest(session, controller, command);
-    }
-
-    @Override
-    public int onSetMediaUri(
-        MediaSession session, ControllerInfo controller, Uri uri, Bundle extras) {
-      assertThat(TextUtils.equals(SUPPORT_APP_PACKAGE_NAME, controller.getPackageName())).isTrue();
-      onSetMediaUriCalled = true;
-      this.uri = uri;
-      this.extras = extras;
-      countDownLatch.countDown();
-      return RESULT_SUCCESS;
     }
 
     @Override